Add basic infrastructure for unit tests
The main logic is is to run all unit tests before generating new html so
this way I can be sure the the chance for any error is as near to 0 as
possible. If any test fails no generation is happening.

So far there are no real unit tests, just a basic infrastructure.

Reason: I do not want to dive into future plans before having a tool for
unit tests, this can be dangerous from now on. So far, it was ok.