Fix potential FileNotFound errors
Let's say that you just cloned the code, run the tests and as a result
you ended up with a bunch of FileNotFound errors because of the missing
/temp folder. In the process this directory is being created so in the
next test all is fine, no errors. This commit fixes this first test scenario.

I have placed a todo for deleting that /temp at the end so this issue
might never happen again undetected "under my radar".

Add unit tests for an uncovered scenario.