[BugFix] Set HTML backup dir as potentially missing
By default the backup directory is not present within the git repository,
so when you try to run your first build the code will generate an error
because there is no backup directory to where to copy the html file.

With this commit if the directory does not exists it will be created
anyway. More elegant solution, less chance to run into error.

Add more unit tests for dirPathChecker.