Refactor uTest.py by renamings
Rename relativeDirPath to dirPath. It does not matter if it is an absolute
or a relative path.

I had named it relative at that time to remember myself to use relative
paths so this way the code might stay independent of where the git repo
were cloned. Now with the filePathType and dirPathType it is not a
requirement to look for anymore.