Further refactor getActualOrPreviousValue (again)
Renamed to getLastValueByFoundEquals because this is way more
self-explanatory and intuitive to use than the previous names.

Fixed a false negative which could return a value looking result within
an another value in some special cases. This automatically fixes the
same issue with other functions as well.

WIP: the function works but it is ugly to read.