Adjust unit tests to be able to run from anywhere
So far I have assumed that the script is run from /webpage/generator.
From now there is freedom to run the script from any location.

Some specialized test cases will just not evaluate if you run the script
outside of the project (e.g. finding the git repo or branch) or outside of
the drive (e.g. calculating relative path from C:/Program files to D:/project)

This behavior will not be a problem since most of the times the script
will be run inside of the project.