Fix potential bug when renaming .git directory as a test
There is a test which renames the directory .git to .gitX. If any file is
already being used by another process than the old file remains there
unremoved leaving a corrupt .git file and a failed test.

This is a quick but dirty solution, not guaranteed to fix the race condition
but will work for now. I will need a safe move function in the future.